Microsoft fixes a few warnings

+++ src/engextcpp/engextcpp.cpp
@@ -223,7 +223,7 @@ ExtCommandDesc::ExtCommandDesc(_In_ PCSTR Name,
 
     if (strlen(Name) > s_LongestCommandName)
     {
-        s_LongestCommandName = strlen(Name);
+        s_LongestCommandName = static_cast<ULONG>(strlen(Name));
     }
 }
 
@@ -1206,7 +1206,7 @@ ExtExtension::OutWrapStr(_In_ PCSTR String)
 {
     if (m_TestWrap)
     {
-        m_TestWrapChars += strlen(String);
+        m_TestWrapChars += static_cast<ULONG>(strlen(String));
         return;
     }
     
@@ -1309,7 +1309,7 @@ ExtExtension::CopyCircleString(_In_ PCSTR Str)
     PSTR Buf;
     ULONG Chars;
     
-    Chars = strlen(Str) + 1;
+    Chars = static_cast<ULONG>(strlen(Str) + 1);
     Buf = RequestCircleString(Chars);
     memcpy(Buf, Str, Chars * sizeof(*Str));
     return Buf;
@@ -1348,7 +1348,7 @@ ExtExtension::SetAppendBuffer(_In_reads_(BufferChars) PSTR Buffer,
 void WINAPI
 ExtExtension::AppendBufferString(_In_ PCSTR Str)
 {
-    ULONG Chars;
+    size_t Chars;
     
     Chars = strlen(Str) + 1;
     if (Chars > m_AppendBufferChars ||
@@ -1361,7 +1361,7 @@ ExtExtension::AppendBufferString(_In_ PCSTR Str)
     memcpy(m_AppendAt, Str, Chars * sizeof(*Str));
     // Position next append where it will overwrite the terminator
     // to continue the existing string.
-    m_AppendAt += Chars - 1;
+    m_AppendAt += static_cast<ULONG>(Chars - 1);
 }
 
 void WINAPI
@@ -1791,7 +1791,7 @@ ExtExtension::CallDebuggeeBase(_In_ PCSTR CommandString,
     ExtCaptureOutputA IgnoreOut;
 
     Cmd.Copy(".call ", 6);
-    Cmd.Append(CommandString, strlen(CommandString) + 1);
+    Cmd.Append(CommandString, static_cast<ULONG>(strlen(CommandString) + 1));
 
     if (FAILED(Status = m_Control->
                Execute(DEBUG_OUTCTL_IGNORE,
@@ -2621,7 +2621,7 @@ ExtExtension::HandleKnownStruct(_In_ PDEBUG_CLIENT Client,
         Status = S_OK;
         while (Struct && Struct->TypeName)
         {
-            ULONG Chars = strlen(Struct->TypeName) + 1;
+            ULONG Chars = static_cast<ULONG>(strlen(Struct->TypeName) + 1);
             CharsNeeded += Chars;
             
             if (Status != S_OK || *BufferChars < Chars)
@@ -2721,7 +2721,7 @@ ExtExtension::HandleQueryValueNames(_In_ PDEBUG_CLIENT Client,
     Status = S_OK;
     while (ExtVal && ExtVal->ValueName)
     {
-        ULONG Chars = wcslen(ExtVal->ValueName) + 1;
+        ULONG Chars = static_cast<ULONG>(wcslen(ExtVal->ValueName) + 1);
         CharsNeeded += Chars;
             
         if (Status != S_OK || BufferChars < Chars)
@@ -3627,7 +3627,7 @@ ExtExtension::HelpAll(void)
     ExtCommandDesc* Desc = m_Commands;
     while (Desc)
     {
-        ULONG NameLen = strlen(Desc->m_Name);
+        ULONG NameLen = static_cast<ULONG>(strlen(Desc->m_Name));
         OutWrap("  !%s%*c- ",
                 Desc->m_Name,
                 m_LongestCommandName - NameLen + 1, ' ');
@@ -4367,7 +4367,7 @@ ExtRemoteTyped::ErtIoctl(_In_ PCSTR Message,
         StrBufferChars * sizeof(*StrBuffer);
     if (InStr)
     {
-        ExtDataBytes += (strlen(InStr) + 1) * sizeof(*InStr);
+        ExtDataBytes += static_cast<ULONG>((strlen(InStr) + 1) * sizeof(*InStr));
     }
 
     ExtData = (EXT_TYPED_DATA*)DataHolder.Get(ExtDataBytes);
